Create styling for asciidoctor-based guides
- Add guide-specific asciidoctor formatting to guide.html
- Add fix-asciidoctor-styles.css to "fix" layout (only in guide.html)
- Ensure asciidoc-specific rules are scoped.
- Pulled in the internet-based prettify so they're served from static

> **Note:** This is a short-term band-aid. The overall css picture needs to be
refactored at some point to account for these fixes.

To ensure we don't break markdown-generated content, or any other content for
that matter.  I also added a big ol' comment to the css to make sure
we remember why we're doing this, and that it is short-term-only.
-- @briancavalier

This was tested against both markdown and asciidoc versions of the same
guide to ensure maximum fidelity was maintained. In fact, more copy buttons
were properly rendered on asciidoctor that were actually missing from markdown.