Refactor most code into self-contained features.
Instead of 2 levels of Javascript granularity, page and component, we
decided to go with a single level for now: feature.  A feature is a
coarse-grained bit of functionality that can be injected into a page.
A template may request a feature by adding a data-<feature-name>
attribute to its <html> element.  The feature bootstrapping
infrastructure in app/main.js will scan the <html> element at startup
time (which doesn't require waiting for domReady), and initialized the
required features.

There are a couple remaining bits that need to be turned into
features: clipboard button handling, sts integration, and the
code-related interactions (code download, etc.) in the guide sidebar.