Add feature for map and newsletter subscription
The newsletter subscription was an inline iframe and was blocking
DOMReady for up to several *seconds*.  Since most of the JS needs to
initialize a lot of stuff after DOMReady, this was holding up
initializion of things that could have been proceeding in
parallel. The newsletterSubscription feature injects the iframe
dynamically so it doesn't block DOMReady.