Register h2 web Console for development profile
Prior to this commit, one couldn't connect to the H2
database in dev mode, since it's in-memory and tied
to its JVM instance (and no TCP port open).

This commit registers a H2 web console servlet within
the same JVM, on "/console/". One can connect to the
database using "jdbc:h2:mem:sagan;MODE=PostgreSQL".
This change is enabled if "development" profile is active.