Split client application in loader + main module
Prior to this commit, the whole client application
was contained in a single file (included with the module
loader).

This change splits the client application in several parts:
- the curl module loader
- the main module "run"

This change triggers an additional HTTP request when loading
the application - but it is now compatible with running the
application locally, from sources (without running the client
build).