Attempt to fix search box timing issue.
This has proven to be impossible to reproduce locally, so I had to
resort to staring at the static code and thinking about what might be
happening.  So this may not work at all :) But it removes an
opportunity for the SearchController's internal state to get out of
sync with the DOM, and simplifies the code in the process.

I think we just have to deploy this on staging or production and see
if we can still reproduce the issue.