Use the latest version of NekoHTML (1.9.20)
The use of self-closing 'a' tags causes a problem in the guides index
page. Self-closing 'a' tags are not legal HTML 5. The only tags which
may be self-closed are void tags. See [1] for more information.

With the upgrade to 1.9.20, the self-closing tag is treated as still
being open and a nested 'a' tag is then encountered. When this tag is
closed, it's actually the self-closing tag which is closed which
causes the stack to unwind prematurely. This causes expression
evaluation to fail (${guide} is null) as the expression context is
incorrect.

The page in question has been updated to replace <a/> with <a></a>.

[1] http://www.w3.org/TR/html5/syntax.html#void-elements