Fix ElasticSearch Health endpoint
Because Sagan is using a high-availability, shared ElasticSearch
instance cluster in production, our credentials don't allow us to issue
"admin level" queries - i.e. all queries related to the cluster state.

So the query used here is fetching index statistics in order to verify
that the service itself is available, and also that we're not going
over quota.