Fix Webhook tests, were platform-dependent
Prior to this commit, the webhook tests were platform-dependent; the
JSON payload is parsed from a file and its content is hashed. Depending
on the platform, the line return chars can change and thus change the
hash string.

This commit removes the line return chars before performing the hash
function.

Fixes #619