Refactor resource handling configuration
Because System.js is a JS module loader, we can't rely anymore on a
single ContentVersionStrategy for all resources.

This commit configures a FixedVersionStrategy for all System.js
resources and derives the application version string from the git
commmit information.

This information is also injected in views via a dedicated helper bean.
This information can't be injected as a `@ModelAttribute`, since
ViewController views do not qualify as Controllers and don't use the
`@ControllerAdvice` infrastructure.