Add API endpoints for /project_metadata
User can GET or delete a release via /{projectId}/{version} or
add/edit a release by POSTing the whole release to /{projectId}.

Authentication is via HTTP Basic with a Github API token (as the
username and an empty password, just like Github native APIs).
All the new endpoints are protected.