Move project info to guide metadata
Prior to this commit, related projects and other tag information was
kept directly within the asciidoctor document as metadata. To get that
information, one needs to fetch and render the full document, which
causes scalability and performance issues.

This commit moves the related project information to the Github
repository description itself, which you can fetch through the Github
API when listing the repositories of an organization.

With that change, the tags list has been replaced by the related
projects information. Tags are still available in the document itself.

Fixes gh-574