Refactor domain model for Projects
This commit completely refactors the domain model for Projects and
Groups, Releases, Samples, etc. This commit also adds a new Generation
concept, which describes the support policy for a given generation of
releases.

These changes are reflected in a new Flyway migration script as well as
the fixtures. Because the latest migration script uses
postgresql-specific queries, the standalone H2 configuration now enables
the postgresql mode.

Fixes gh-903