Remove Project Metadata API
Prior to this commit, this application would expose a Web API under
`"/project_metadata/*"`. This was initially used for external project
pages hosted on projects.spring.io (previously, GitHub pages).
It's also been used by start.spring.io and Initializr instances to fetch
the list of available Spring Boot versions at runtime.

This is problematic since this Web API is merely serializing the raw
domain model as JSON. Any change in the model results in incompatible
changes in the API.

Now that all project pages haven been moved to this application, we can
focus on the remaining use case and ship tailored JSON payloads to
Initializr instances until they upgrade to the new API.

This commit removes the former implementation and provides a legacy
endpoint for Initializr instances only.

Closes gh-978